requirements.txt
streamlit>=1.36
python-dotenv>=1.0
openai>=1.40
python-docx>=1.1
requests>=2.31